+++ fj-daogen-base/src/main/java/org/fugerit/java/daogen/base/gen/DaogenBasicGenerator.java
@@ -33,6 +33,8 @@ public abstract class DaogenBasicGenerator extends SimpleJavaGenerator implement
 	
 	protected static final String JD_THROWS = " * @throws ";
 	
+	protected static final String CLOSE_AND_THROWS = " ) throws ";
+	
 	protected static final String AT_OVERRIDE = "@Override";
 	
 	protected static final String REAL_CLASS_COMMENT = TAB+"// [HELPER/IMPL MODEL] this class is a stub and can be modified as you see fit (it will not been overwritten)";